+++ src/containers/NotificationModalContainer.tsx
@@ -1,8 +1,7 @@
 import { useEffect, useState, useMemo } from "react";
-import Close from "/public/icons/ic_close.svg";
-import Dot from "/public/icons/ic_dot.svg";
 import { getNotifications, deleteNotifications } from "@/api/notification";
-import { timeDiff, getDotColor } from "@/utils/timeDiff";
+import ModalOverlay from "@/components/ui/Modal/ModalOverlay";
+import NotificationModal from "@/components/NotificationModal";
 
 interface Notification {
   createdAt: string;
@@ -10,91 +9,70 @@ interface Notification {
   id: number;
 }
 
-interface NotificationModalProps {
-  createdAt: string;
-  content: string;
-  id: number;
-  onDelete: (id: number) => void;
+interface ModalProps {
+  isOpen: boolean;
+  onClose: () => void;
 }
 
-const NotificationModal = ({
-  createdAt,
-  content,
-  id,
-  onDelete,
-}: NotificationModalProps) => {
-  const timeText = useMemo(() => timeDiff(createdAt), [createdAt]);
-  const dotColor = useMemo(() => getDotColor(timeText), [timeText]);
-
-  const handleDelete = async () => {
-    await deleteNotifications(id);
-    onDelete(id);
-  };
-
-  return (
-    <div className="w-[328px] px-3 py-4 mb-2 bg-background border border-green-50 rounded-[5px]">
-      <div className=" flex justify-between">
-        <Dot className={dotColor} />
-        <button type="button" onClick={handleDelete}>
-          <Close width={24} height={24} />
-        </button>
-      </div>
-      <div className="mb-1 text-[14px] leading-[22px] text-notic-text">
-        {content}
-      </div>
-      <div className="text-[12px] leading-[16px] text-notice-gray-2">
-        {timeDiff(createdAt)}
-      </div>
-    </div>
-  );
-};
-
-const NotificatonModalContainer = () => {
+const NotificatonModalContainer = ({ isOpen, onClose }: ModalProps) => {
   const [totalCount, setTotalCount] = useState<number>(0);
   const [notifications, setNotifications] = useState<Notification[]>([]);
-
 
   useEffect(() => {
     const fetchNotifications = async () => {
       const res = await getNotifications({ page: 1, pageSize: 10 });
-      console.log(res)
+      console.log(res);
       setTotalCount(res.totalCount);
       setNotifications(res.list);
     };
+
     fetchNotifications();
-  }, []);
+  }, [isOpen]);
 
   const handleDelete = (id: number) => {
     setNotifications((prev) =>
       prev.filter((notification) => notification.id !== id)
     );
+    setTotalCount((prev) => prev - 1);
   };
 
-  if (totalCount === 0) {
-    return (
+  return totalCount === 0 ? (
+    <ModalOverlay
+      isOpen={isOpen}
+      onClose={onClose}
+      bgColor="bg-notice-bg"
+      closeButtonSize="w-6 h-6"
+      closeButtonColor="text-notice-text"
+    >
       <div className="w-[328px] mt-10 px-3 py-4 font-semibold bg-background border border-green-50 rounded-[5px]">
         새로운 알림이 없습니다.
       </div>
-    );
-  }
-
-  return (
-    <>
+    </ModalOverlay>
+  ) : (
+    <ModalOverlay
+      isOpen={isOpen}
+      onClose={onClose}
+      bgColor="bg-notice-bg"
+      closeButtonSize="w-6 h-6"
+      closeButtonColor="text-notice-text"
+    >
       <div>
         <div className="mt-1 text-[25px] leading-[25.04px] font-bold text-notice-text mb-4">
           {`알림${totalCount}개`}
         </div>
-        {notifications.map((notification) => (
-          <NotificationModal
-            key={notification.id}
-            id={notification.id}
-            createdAt={notification.createdAt}
-            content={notification.content}
-            onDelete={handleDelete}
-          />
-        ))}
+        <ul>
+          {notifications.map((notification) => (
+            <NotificationModal
+              key={notification.id}
+              id={notification.id}
+              createdAt={notification.createdAt}
+              content={notification.content}
+              onDelete={handleDelete}
+            />
+          ))}
+        </ul>
       </div>
-    </>
+    </ModalOverlay>
   );
 };
